backend/requirements.txt
# API
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
httpx==0.27.2

# Vector DB
qdrant-client==1.11.3

# Document Extraction
PyMuPDF==1.24.10
openpyxl==3.1.5
python-docx==1.1.2
python-pptx==1.0.2
beautifulsoup4==4.12.3
lxml==5.3.0

# Embeddings
torch>=2.0.0
sentence-transformers==3.1.1

# Keyword Extraction (reemplaza al LLM generativo)
keybert==0.8.5

# LLM Providers
google-generativeai>=0.8.0

# Config
pydantic-settings==2.5.2
python-dotenv==1.0.1
pytesseract
pdf2image

# Auth
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
